refactor: put parsing scripts in new parser implementation
